renamed folder from router to mux
